+++ homework02/src/main/java/ru/otus/asamofalov/dao/QuestionDaoCsv.java
@@ -7,7 +7,7 @@ import ru.otus.asamofalov.domain.ChosenAnswerQuestion;
 import ru.otus.asamofalov.domain.FreeAnswerQuestion;
 import ru.otus.asamofalov.domain.Question;
 import ru.otus.asamofalov.exception.DaoException;
-import ru.otus.asamofalov.helper.CsvParser;
+import ru.otus.asamofalov.service.CsvParser;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -17,9 +17,11 @@ import java.util.List;
 public class QuestionDaoCsv implements QuestionDao {
 
     private final String fileName;
+    private final CsvParser csvParser;
 
-    public QuestionDaoCsv(@Value("${filename}") String fileName) {
+    public QuestionDaoCsv(@Value("${filename}") String fileName, CsvParser csvParser) {
         this.fileName = fileName;
+        this.csvParser = csvParser;
     }
 
     @Override
@@ -28,10 +30,10 @@ public class QuestionDaoCsv implements QuestionDao {
         List<Question> questions = new ArrayList<>();
 
         try {
-            for (String[] row : CsvParser.parse(fileName)) {
+            for (String[] row : csvParser.parse(fileName)) {
 
                 if (row.length > 1 && row[1].contains(Constants.ANSWERS_DELIMITER)) {
-                    questions.add(new ChosenAnswerQuestion(row[0], row[2], row[1]));
+                    questions.add(new ChosenAnswerQuestion(row[0], row[2], row[1].split(Constants.ANSWERS_DELIMITER)));
                 } else {
                     questions.add(new FreeAnswerQuestion(row[0], row[2]));
                 }
