+++ homework02/src/main/java/ru/otus/asamofalov/exception/AppException.java
@@ -1,6 +1,6 @@
 package ru.otus.asamofalov.exception;
 
-public class AppException extends Exception {
+public class AppException extends RuntimeException {
 
     public AppException() {
         super();
