+++ homework02/src/main/java/ru/otus/asamofalov/service/CsvParser.java
@@ -1,0 +1,7 @@
+package ru.otus.asamofalov.service;
+
+import java.io.IOException;
+
+public interface CsvParser {
+    String[][] parse(String fileName) throws IOException;
+}
